CRAG/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langgraph
faiss-cpu
tavily-python
python-dotenv
pydantic
